Separating out the streamer into its own file. Cleaning up and adding some callbacks etc.
